Moved the Orchestration.OrchestrationInit class to the Init subfolder.
