+++ WonkaSystem/WonkaEth/WonkaEth.csproj
@@ -97,6 +97,7 @@
     <Compile Include="Extensions\WonkaGroveExtensions.cs" />
     <Compile Include="Contracts\ISerialize.cs" />
     <Compile Include="Contracts\IOrchestrate.cs" />
+    <Compile Include="Orchestration\Init\OrchestrationInitData.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
@@ -108,6 +109,9 @@
     <Folder Include="Misc\" />
     <Folder Include="Orchestration\" />
     <Folder Include="Init\" />
+    <Folder Include="Orchestration\BlockchainEvents\" />
+    <Folder Include="Orchestration\BlockchainOutput\" />
+    <Folder Include="Orchestration\Init\" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\WonkaBre\WonkaBre.csproj">
